LPS-93010 Move to portal-security-permission-test
