+++ modules/apps/portal-security/portal-security-permission-test/src/testIntegration/resources/com/liferay/portal/security/permission/test/dependencies/resource-actions.xml
